- Consolidates code in /server/auth, placing it in /server/users; everything updated and appears to work just fine; - closes issue #105;
<3,
Vita
