+++ server/middleware/file-upload.js
@@ -1,0 +1,5 @@
+const multer = require('multer');
+
+const fileUpload = multer({});
+
+module.exports = fileUpload;
