+++ server/package.json
@@ -25,6 +25,7 @@
     "mongodb": "^3.6.6",
     "mongoose": "^5.12.7",
     "mongoose-unique-validator": "^2.0.3",
+    "multer": "^1.4.3",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
