--- server/routes/auth-routes.js
@@ -1,20 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const { check } = require('express-validator');
-const authControllers = require('../controllers/auth-controllers');
-
-// Logging in
-router.post(
-  '/login',
-  check('username')
-    .not()
-    .isEmpty(),
-  check('password')
-    .not()
-    .isEmpty(),
-  authControllers.login);
-
-// Logging out
-router.post('/logout', authControllers.logout);
-
-module.exports = router;
